+++ frontend/projects/valtimo/admin-settings/src/lib/models/index.ts
@@ -16,3 +16,4 @@
 
 export * from './accent-colors.model';
 export * from './feature-toggle.model';
+export * from './menu-configuration.model';
