+++ frontend/projects/valtimo/admin-settings/src/lib/models/menu-configuration.model.ts
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2015-2026 Ritense BV, the Netherlands.
+ *
+ * Licensed under EUPL, Version 1.2 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {MenuConfigurationItem} from '@valtimo/components';
+import {ExternalPluginMenuPage} from '@valtimo/plugin';
+
+/** A draggable entry in the right-hand "Available items" palette. */
+interface PaletteItem {
+  paletteType: 'catalog' | 'group' | 'section-header' | 'custom-link' | 'plugin-page';
+  label: string;
+  icon?: string;
+  itemId?: string;
+  page?: ExternalPluginMenuPage;
+}
+
+interface PaletteGroup {
+  categoryKey: string;
+  items: PaletteItem[];
+}
+
+/** Builder node = a stored config node plus a transient `_uid` for stable identity across redraws. */
+type BuilderNode = MenuConfigurationItem & {_uid: string; children?: BuilderNode[]};
+
+export {PaletteItem, PaletteGroup, BuilderNode};
